feat: refactored typrography comp a little bit
extending the props of chakra typography and adding more props.
since we are extending that, wee dont need fontweight, fontfamily etc expli
citely defined at our side as it is coming with more extra props in the propsmodel
of chakra
